halflife.py: switch to API v2, revive domain search
msapi.py: split out Metasmoke API to a separate module

This is a quick and dirty refactoring but it restores some of the
functionality which was dsiabled when Metasmoke regex search was
turned off for performance and stability reasons.

domain search is now examining the domain records. Paging means we
still get partial results for domains with many hits, but it could
be construed as a feature rather than a bug.